Move aarch64 relocation code and associated types to linker-utils
This will allow some of this code to be reused by linker-diff.
