+++ libwild/src/arch.rs
@@ -1,11 +1,11 @@
 //! Abstraction over different CPU architectures.
 
 use crate::args::OutputKind;
-use crate::elf::RelocationKindInfo;
 use crate::error::Result;
 use crate::resolution::ValueFlags;
 use anyhow::bail;
 use linker_utils::elf::DynamicRelocationKind;
+use linker_utils::elf::RelocationKindInfo;
 use linker_utils::elf::SectionFlags;
 use linker_utils::relaxation::RelocationModifier;
 use std::borrow::Cow;
@@ -63,7 +63,7 @@ pub(crate) trait Relaxation {
 
     fn apply(&self, section_bytes: &mut [u8], offset_in_section: &mut u64, addend: &mut i64);
 
-    fn rel_info(&self) -> crate::elf::RelocationKindInfo;
+    fn rel_info(&self) -> RelocationKindInfo;
 
     fn debug_kind(&self) -> impl std::fmt::Debug;
